+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project follows semantic versioning.
 
 ### Unpublished
     - [fixed] Bug in `Array` division.
+    - [fixed] Bug where `Rem` would sometimes exit early with the wrong answer.
+    - [added] `PartialDiv` operator that performs division as a partial function -- it's defined
+      only when there is no remainder.
 
 ### 1.5.2 (2017-02-04)
     - [fixed] Bug between `Div` implementation and type system.